Fix MSVC warning

+++ src/lightmap.cpp
@@ -375,8 +375,8 @@ void map::generate_lightmap( const int zlev )
                         if( light_emitted > 0 ) {
                             add_light_source( p, light_emitted );
                         }
-                        const int light_override = cur->local_light_override();
-                        if( light_override >= 0.0 ) {
+                        const float light_override = cur->local_light_override();
+                        if( light_override >= 0.0f ) {
                             lm_override.push_back( std::pair<tripoint, float>( p, light_override ) );
                         }
                     }
